Scan / Swap Jewel Logic
A base line of code that I need to expand on for match verification.

I can scan the grid and swap a jewel with this, however I still need to finalise the match 3+ for verification / score.

Line 61/83 is just a test / thought out start of how you verify by getting the jewel and seeing if the kind matches up